+++ src/apps/beefy/common/beefy.vault-token-fetcher.ts
@@ -1,5 +1,4 @@
 import { Inject } from '@nestjs/common';
-import { Contract } from 'ethers';
 
 import { APP_TOOLKIT, IAppToolkit } from '~app-toolkit/app-toolkit.interface';
 import {
@@ -9,7 +8,7 @@ import {
   UnderlyingTokensStageParams,
 } from '~position/template/app-token.template.position-fetcher';
 
-import { BeefyContractFactory } from '../contracts';
+import { BeefyContractFactory, BeefyVaultToken } from '../contracts';
 
 import { BeefyVaultTokenDefinitionsResolver } from './beefy.vault.token-definition-resolver';
 
@@ -17,8 +16,8 @@ export type BeefyVaultTokenDataProps = {
   liquidity: number;
 };
 
-export abstract class BeefyVaultTokenFetcher<T extends Contract> extends AppTokenTemplatePositionFetcher<
-  T,
+export abstract class BeefyVaultTokenFetcher extends AppTokenTemplatePositionFetcher<
+  BeefyVaultToken,
   BeefyVaultTokenDataProps
 > {
   constructor(
@@ -30,35 +29,34 @@ export abstract class BeefyVaultTokenFetcher<T extends Contract> extends AppToke
     super(appToolkit);
   }
 
-  private getVaultDefinitions() {
-    return this.tokenDefinitionsResolver.getVaultDefinitions(this.network);
-  }
-
-  protected async selectVault(vaultAddress: string) {
-    const vaultDefinitions = await this.getVaultDefinitions();
-    return vaultDefinitions.find(v => v.address.toLowerCase() === vaultAddress) ?? null;
+  getContract(address: string): BeefyVaultToken {
+    return this.contractFactory.beefyVaultToken({ network: this.network, address });
   }
 
   async getAddresses(): Promise<string[]> {
-    const vaultDefinitions = await this.getVaultDefinitions();
+    const vaultDefinitions = await this.tokenDefinitionsResolver.getVaultDefinitions(this.network);
     return vaultDefinitions.map(({ address }) => address.toLowerCase());
   }
 
-  async getUnderlyingTokenAddresses({ contract }: UnderlyingTokensStageParams<T>): Promise<string[]> {
-    const vault = await this.selectVault(contract.address.toLowerCase());
+  async getUnderlyingTokenAddresses({ address }: UnderlyingTokensStageParams<BeefyVaultToken>): Promise<string[]> {
+    const vaultDefinitions = await this.tokenDefinitionsResolver.getVaultDefinitions(this.network);
+    const vault = vaultDefinitions.find(v => v.address.toLowerCase() === address) ?? null;
     if (!vault) throw new Error('Cannot find specified vault');
 
     return [vault.underlyingAddress.toLowerCase()];
   }
 
-  async getPricePerShare({ contract, appToken, multicall }: PricePerShareStageParams<T>): Promise<number | number[]> {
-    const ratioRaw = await multicall.wrap(contract).getPricePerFullShare();
+  async getPricePerShare({
+    contract,
+    appToken,
+  }: PricePerShareStageParams<BeefyVaultToken>): Promise<number | number[]> {
+    const ratioRaw = contract.getPricePerFullShare();
     const decimals = appToken.decimals;
 
     return Number(ratioRaw) / 10 ** decimals;
   }
 
-  async getDataProps(opts: DataPropsStageParams<T, BeefyVaultTokenDataProps>) {
+  async getDataProps(opts: DataPropsStageParams<BeefyVaultToken, BeefyVaultTokenDataProps>) {
     const { appToken } = opts;
     const reserve = Number(appToken.pricePerShare) * appToken.supply;
     const liquidity = reserve * appToken.tokens[0].price;