+++ src/apps/tectonic/tectonic.module.ts
@@ -6,6 +6,7 @@ import { TectonicContractFactory } from './contracts';
 import { CronosTectonicBalanceFetcher } from './cronos/tectonic.balance-fetcher';
 import { CronosTectonicBorrowContractPositionFetcher } from './cronos/tectonic.borrow.contract-position-fetcher';
 import { CronosTectonicSupplyTokenFetcher } from './cronos/tectonic.supply.token-fetcher';
+import { CronosTectonicXTonicTokenFetcher } from './cronos/tectonic.xtonic.token-fetcher';
 import { TectonicBorrowBalanceHelper } from './helper/tectonic.borrow.balance-helper';
 import { TectonicClaimableBalanceHelper } from './helper/tectonic.claimable.balance-helper';
 import { TectonicSupplyBalanceHelper } from './helper/tectonic.supply.balance-helper';
@@ -19,13 +20,13 @@ import { TectonicAppDefinition, TECTONIC_DEFINITION } from './tectonic.definitio
     CronosTectonicBalanceFetcher,
     CronosTectonicBorrowContractPositionFetcher,
     CronosTectonicSupplyTokenFetcher,
+    CronosTectonicXTonicTokenFetcher,
     TectonicAppDefinition,
-    TectonicContractFactory,
-    TectonicClaimableBalanceHelper,
     TectonicBorrowBalanceHelper,
-    TectonicSupplyTokenHelper,
-    TectonicSupplyBalanceHelper,
+    TectonicClaimableBalanceHelper,
     TectonicContractFactory,
+    TectonicSupplyBalanceHelper,
+    TectonicSupplyTokenHelper,
   ],
   exports: [
     TectonicClaimableBalanceHelper,